feat: agregar paginación y manejo de errores en el servicio de usuarios

+++ src/ranking/entities/ranking.entity.ts
@@ -5,7 +5,7 @@ import { Column, Entity, JoinColumn, OneToOne} from "typeorm";
 @Entity()
 export class Ranking extends BaseEntity {
 
-    @Column({type:"numeric", nullable: false})
+    @Column({type:"int", nullable: false})
     position:number;
 
     @Column({type:"numeric", nullable: false})
